Setup git submodules & an rcrc
Antigen is a dependency I have 100% so its now a submodule
of this repo so automatically have it.

Also setup rcrc so certain files are ignored when rcm is run/setup
because shit like readme & terminal theme don't and shouldn't be
symlinked nor should i have to remember i have that shit.